requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-filter==23.5
django-guardian==2.4.0
django-storages==1.14.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
celery==5.3.6
redis==5.0.1
Pillow==10.1.0
reportlab==4.0.8
django-simple-history==3.4.0
django-audit-log==0.7.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
django-address==0.2.8
django-money==3.4.1
django-notifications-hq==1.8.2
python-dateutil==2.8.2
requests==2.31.0
PyJWT==2.8.0
pydantic==2.5.2
pydantic[email]  # For EmailStr support
pydantic-settings==2.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0
mysqlclient==2.2.1
django-mysql==4.12.0
